Create First Brainteaser
As a new student coming to aptech the the first thing we look at was making reference to html 5 structure,starting with "doctype  html "human consciousness must always be there. Following the principles of human consciousness make reference to three different tags of your choice, These tags must contain construct that represents the value of the browser.