Brainteaser3
In a section for a school pattern such as aptech we have structured and unstructured codes. The styler follows a certain html standard . At a certain point in time in the header tag was misinterpreted by the styler who failed to follow the html standard. The html file consists of 6 different tags that follow the default html standard as a styler deduce the context above.